precommit

+++ src/find_fills.py
@@ -1,9 +1,10 @@
-from typing import Set, List
-from data_types import Team, Player, PlayerAssignment, clip_value, PlayerRole
 import math
+from typing import List, Set
+
+from data_types import Player, PlayerRole, Team, clip_value
 
 
-def find_fills(team: Team, all_players: Set[Player], ideal_score: int) -> List[str]:
+def find_fills(team: Team, all_players: Set[Player], ideal_score: float) -> List[str]:
     """Find a fill, where we're aiming to hit the average score of finalized teams.
 
     The fill is generally a range, and any kind of player can fill. Players who are primarily queens will have
